separate processes from data transfers
