+++ tests/test_on_data/test_yaml_conversion_specification_cli.py
@@ -5,7 +5,7 @@ from pathlib import Path
 from hdmf.testing import TestCase
 from pynwb import NWBHDF5IO
 
-from neuroconv.tools.data_transfers import deploy_process
+from neuroconv.tools import deploy_process
 
 from .setup_paths import ECEPHY_DATA_PATH as DATA_PATH
 from .setup_paths import OUTPUT_PATH
